backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
whitenoise>=6.6
gunicorn>=21.0
psycopg2-binary>=2.9
